requirements.txt
datasets
torch
open_clip_torch
matplotlib
seaborn
pandas
numpy
scipy
ultralytics
opencv-python
pyserial
pyyaml
